+++ frontend/src/redux/actions/actionTypes.js
@@ -1,0 +1,16 @@
+export const actionTypes = {
+    GET_RECIPE: 'recipe/getRecipe',
+    LOGIN: 'user/login',
+    ADD_FAVOURITE_RECIPE: 'user/addFavouriteRecipe',
+    DELETE_FAVOURITE_RECIPE: 'user/deleteFavouriteRecipe',
+    GET_FAVOURITE_RECIPES: 'user/getFavouriteRecipes',
+    DISPLAY_RECIPE: 'recipe/displayRecipe'
+}
+
+
+
+
+
+
+
+
