+++ frontend/src/redux/actions/actions.js
@@ -1,13 +1,7 @@
 import { setUsername, setUserImage, addRecipe } from "redux/reducers/userReducer";
-
-export const actionTypes = {
-    GET_RECIPE: 'recipe/getRecipe',
-    LOGIN: 'user/login',
-    ADD_FAVOURITE_RECIPE: 'user/addFavouriteRecipe',
-    DELETE_FAVOURITE_RECIPE: 'user/deleteFavouriteRecipe',
-    GET_FAVOURITE_RECIPES: 'user/getFavouriteRecipes'
-}
+import { displayRecipe } from "redux/reducers/recipeReducer";
 
 export const SET_USERNAME = setUsername;
 export const SET_USER_IMAGE = setUserImage;
 export const ADD_RECIPE = addRecipe;
+export const DISPLAY_RECIPE = displayRecipe;
